extract task-queue to a new file
